+++ python/llama-index-server/examples/simple-agent/app/workflow.py
@@ -3,7 +3,7 @@ from typing import Optional
 from llama_index.core.agent.workflow import AgentWorkflow
 from llama_index.core.settings import Settings
 from llama_index.llms.openai import OpenAI
-from llama_index.server.api.models import ChatRequest
+from llama_index.server.models import ChatRequest
 
 
 def create_workflow(chat_request: Optional[ChatRequest] = None) -> AgentWorkflow:
