+++ python/llama-index-server/llama_index/server/__init__.py
@@ -1,4 +1,4 @@
-from .api.models import UIEvent
+from .models.ui import UIEvent
 from .server import LlamaIndexServer, UIConfig
 
 __all__ = ["LlamaIndexServer", "UIConfig", "UIEvent"]
