+++ python/llama-index-server/llama_index/server/api/models.py
@@ -1,265 +1,2 @@
-import logging
-import os
-import re
-from enum import Enum
-from typing import Any, Dict, List, Literal, Optional, Type, Union
-
-from llama_index.core.schema import NodeWithScore
-from llama_index.core.types import ChatMessage, MessageRole
-from llama_index.core.workflow import Event
-from llama_index.core.workflow.events import (
-    HumanResponseEvent as FrameworkHumanResponseEvent,
-)
-from llama_index.core.workflow.events import (
-    InputRequiredEvent,
-)
-from llama_index.server.settings import server_settings
-from llama_index.server.utils import llamacloud
-from pydantic import BaseModel, Field, field_validator
-
-logger = logging.getLogger("uvicorn")
-
-
-class ChatAPIMessage(BaseModel):
-    role: MessageRole
-    content: str
-    annotations: Optional[List[Any]] = None
-
-    def to_llamaindex_message(self) -> ChatMessage:
-        return ChatMessage(role=self.role, content=self.content)
-
-    @property
-    def human_response(self) -> Optional[Any]:
-        if self.annotations:
-            for annotation in self.annotations:
-                if (
-                    isinstance(annotation, dict)
-                    and annotation.get("type") == "human_response"
-                ):
-                    return annotation.get("data", {})
-        return None
-
-
-class ChatRequest(BaseModel):
-    id: str  # see https://ai-sdk.dev/docs/reference/ai-sdk-ui/use-chat#id - constant for the same chat session
-    messages: List[ChatAPIMessage]
-    data: Optional[Any] = None
-
-    @field_validator("messages")
-    def validate_messages(cls, v: List[ChatAPIMessage]) -> List[ChatAPIMessage]:
-        if v[-1].role != MessageRole.USER:
-            raise ValueError("Last message must be from user")
-        return v
-
-    @field_validator("id")
-    def validate_id(cls, v: str) -> str:
-        if re.search(r"[^a-zA-Z0-9_-]", v):
-            raise ValueError("ID contains special characters")
-        return v
-
-
-class AgentRunEventType(Enum):
-    TEXT = "text"
-    PROGRESS = "progress"
-
-
-class AgentRunEvent(Event):
-    name: str
-    msg: str
-    event_type: AgentRunEventType = AgentRunEventType.TEXT
-    data: Optional[dict] = None
-
-    def to_response(self) -> dict:
-        return {
-            "type": "agent",
-            "data": {
-                "agent": self.name,
-                "type": self.event_type.value,
-                "text": self.msg,
-                "data": self.data,
-            },
-        }
-
-
-class SourceNodesEvent(Event):
-    nodes: List[NodeWithScore]
-
-    def to_response(self) -> dict:
-        return {
-            "type": "sources",
-            "data": {
-                "nodes": [
-                    SourceNodes.from_source_node(node).model_dump()
-                    for node in self.nodes
-                ]
-            },
-        }
-
-
-class SourceNodes(BaseModel):
-    id: str
-    metadata: Dict[str, Any]
-    score: Optional[float]
-    text: str
-    url: Optional[str]
-
-    @classmethod
-    def from_source_node(cls, source_node: NodeWithScore) -> "SourceNodes":
-        metadata = source_node.node.metadata
-        url = cls.get_url_from_metadata(metadata)
-
-        return cls(
-            id=source_node.node.node_id,
-            metadata=metadata,
-            score=source_node.score,
-            text=source_node.node.text,  # type: ignore
-            url=url,
-        )
-
-    @classmethod
-    def get_url_from_metadata(
-        cls,
-        metadata: Dict[str, Any],
-        data_dir: Optional[str] = None,
-    ) -> Optional[str]:
-        url_prefix = server_settings.file_server_url_prefix
-        if data_dir is None:
-            data_dir = "data"
-        file_name = metadata.get("file_name")
-
-        if file_name and url_prefix:
-            if llamacloud.is_llamacloud_file(metadata):
-                file_name = llamacloud.get_local_file_name(metadata)
-                return f"{url_prefix}/output/llamacloud/{file_name}"
-            is_private = metadata.get("private", "false") == "true"
-            if is_private:
-                # file is a private upload
-                return f"{url_prefix}/output/uploaded/{file_name}"
-            # file is from calling the 'generate' script
-            # Get the relative path of file_path to data_dir
-            file_path = metadata.get("file_path")
-            data_dir = os.path.abspath(data_dir)
-            if file_path and data_dir:
-                relative_path = os.path.relpath(file_path, data_dir)
-                return f"{url_prefix}/data/{relative_path}"
-        # fallback to URL in metadata (e.g. for websites)
-        return metadata.get("URL")
-
-    @classmethod
-    def from_source_nodes(
-        cls, source_nodes: List[NodeWithScore]
-    ) -> List["SourceNodes"]:
-        return [cls.from_source_node(node) for node in source_nodes]
-
-
-class ComponentDefinition(BaseModel):
-    type: str
-    code: str
-    filename: str
-
-
-class UIEvent(Event):
-    type: str
-    data: BaseModel
-
-    def to_response(self) -> dict:
-        return {
-            "type": self.type,
-            "data": self.data.model_dump(),
-        }
-
-
-class ArtifactType(str, Enum):
-    CODE = "code"
-    DOCUMENT = "document"
-
-
-class CodeArtifactData(BaseModel):
-    file_name: str
-    code: str
-    language: str
-
-
-class DocumentArtifactData(BaseModel):
-    title: str
-    content: str
-    type: Literal["markdown", "html"]
-
-
-class Artifact(BaseModel):
-    created_at: Optional[int] = None
-    type: ArtifactType
-    data: Union[CodeArtifactData, DocumentArtifactData]
-
-    @classmethod
-    def from_message(cls, message: ChatAPIMessage) -> Optional["Artifact"]:
-        if not message.annotations or not isinstance(message.annotations, list):
-            return None
-
-        for annotation in message.annotations:
-            if isinstance(annotation, dict) and annotation.get("type") == "artifact":
-                try:
-                    artifact = cls.model_validate(annotation.get("data"))
-                    return artifact
-                except Exception as e:
-                    logger.warning(
-                        f"Failed to parse artifact from annotation: {annotation}. Error: {e}"
-                    )
-
-        return None
-
-
-class ArtifactEvent(Event):
-    type: str = "artifact"
-    data: Artifact
-
-    def to_response(self) -> dict:
-        return {
-            "type": self.type,
-            "data": self.data.model_dump(),
-        }
-
-
-# TODO: Some of these models should be at a higher level?
-class HumanInputEvent(InputRequiredEvent):
-    """
-    Use this event to request input from a human.
-    It will block the workflow execution until the human responds.
-    """
-
-    response_event_type: Type[Event] = Field(
-        description="The type of event that the workflow is waiting for.",
-    )
-    event_type: str = Field(
-        description="An identifier for the UI component that will be used to render the input.",
-    )
-    data: Union[Dict[str, Any], BaseModel] = Field(
-        description="The data to be sent to the UI component that will be used to render the input.",
-    )
-
-    def __init__(self, **kwargs: Any) -> None:
-        # Construct the prefix for InputRequiredEvent
-        event_type = kwargs.get("event_type", None)
-        data = kwargs.get("data", None)
-        if "prefix" not in kwargs:
-            kwargs["prefix"] = f"Need input for {event_type} with data: {data}"
-        super().__init__(**kwargs)
-
-    def to_response(self) -> dict:
-        return {
-            "type": self.event_type,
-            "data": self.data
-            if isinstance(self.data, dict)
-            else self.data.model_dump(),
-        }
-
-
-class HumanResponseEvent(FrameworkHumanResponseEvent):  # type: ignore
-    """
-    Use this event to send a response from a human.
-    """
-
-    def __init__(self, **kwargs: Any) -> None:
-        if "response" not in kwargs:
-            kwargs["response"] = f"Human response with data: {kwargs.get('data', {})}"
-        super().__init__(**kwargs)
+# TODO: For backward compatibility, remove this in a minor release
+from llama_index.server.models import *  # noqa
