+++ python/llama-index-server/tests/api/test_event_stream.py
@@ -9,9 +9,9 @@ from llama_index.core.agent.workflow.workflow_events import AgentStream
 from llama_index.core.types import MessageRole
 from llama_index.core.workflow import StopEvent
 from llama_index.core.workflow.handler import WorkflowHandler
-from llama_index.server.api.models import ChatAPIMessage, ChatRequest
 from llama_index.server.api.routers.chat import _stream_content
 from llama_index.server.api.utils.vercel_stream import VercelStreamResponse
+from llama_index.server.models.chat import ChatAPIMessage, ChatRequest
 
 
 @pytest.fixture()
